Splittet ut og lagt til tabell for å vise vedtakshistorikk for barnetilsyn
